init match restru
